The velocity send to through emotions has been corrected. The problem was the use of the functin abs, which returns an integer.
The basic test on some changes in emotion was done in the action torso.

There is a first implementation to move keepon's body.

The firmeware for arduinodue has been updated. It now has the new changes in the messages. Additionally, it was added the capability to send servos' information and receive servos' information in an independent way, using a boolean flag to determine which parameter has been changed.

TODO:
	*Add the emotion capabilities for the action keepon_body
	*Implement the controller for TriskarSmall