It was solved a problem with the emotion synchornization.
It was added the possibility to send synch messages from do_nothing action. Thus two new parameters could be added to this actions: emotional_repeat and emotional_time. The first tells if should generate messages more than once, and the second the time between messages.

One problem with the current implementation is that only can have one instance of one action at the time. This could be solved if to each action is added the context that belongs to. However, changes in the nodes to support two or more instances should be added.

TODO:
	*Look how to add multiple instances of the same actions (this makes sense, at least for the moment, just for do_nothing)
	*Implement the emotional changes in velocity for move's Keepon body
	*Implement all the controllers for the triskar small